+++ package.json
@@ -6,6 +6,10 @@
     "publish-all": "lerna publish --conventional-commits --no-private --yes --create-release github --no-changelog",
     "lint": "eslint . --fix --report-unused-disable-directives",
     "lint-no-fix": "eslint . --quiet --report-unused-disable-directives",
+    "test-typings": "yarn workspaces run test-typings",
+    "test-unit": "yarn test-unit-esm && yarn workspaces run test-unit",
+    "test-unit-esm": "mocha esm-named-exports.test.js",
+    "build": "yarn workspaces run build",
     "docs": "typedoc",
     "fix-commit": "export TMPFILE=$(mktemp) && grep -v '^#' $(git rev-parse --git-dir)/COMMIT_EDITMSG > $TMPFILE && git commit -e -F $TMPFILE",
     "------------------------------------- local test dbs --------------------------------------": "",
@@ -43,13 +47,19 @@
     "@ephys/eslint-config-typescript": "18.0.1",
     "@rushstack/eslint-patch": "1.2.0",
     "chai": "4.3.7",
+    "esbuild": "0.17.0",
     "eslint": "8.30.0",
     "eslint-plugin-jsdoc": "39.6.4",
     "eslint-plugin-mocha": "10.1.0",
+    "fast-glob": "3.2.12",
     "lerna": "6.4.1",
     "lint-staged": "13.1.0",
     "markdownlint-cli": "0.32.2",
-    "typedoc": "0.23.23"
+    "mocha": "10.2.0",
+    "node-hook": "1.0.0",
+    "source-map-support": "0.5.21",
+    "typedoc": "0.23.23",
+    "typescript": "4.9.4"
   },
   "packageManager": "yarn@1.22.19",
   "lint-staged": {
