+++ packages/validator-js/.eslintrc.js
@@ -1,0 +1,11 @@
+module.exports = {
+  parserOptions: {
+    project: [`${__dirname}/tsconfig.json`],
+  },
+  overrides: [{
+    files: ['test/**/*'],
+    parserOptions: {
+      project: [`${__dirname}/test/tsconfig.json`],
+    },
+  }],
+};
