--- packages/validator-js/index.js
@@ -1,1 +1,0 @@
-throw new Error('This package does not do anything yet. See https://github.com/sequelize/sequelize/issues/15497');
